requirements.txt

        fastapi==0.109.0
        uvicorn==0.25.0
        requests==2.31.0
        pandas==2.2.0
        numpy==1.26.0
        google-api-python-client==2.108.0
        facebook-business==18.0.0
        twilio==8.10.0
        python-dotenv==1.0.0
        